requirements.txt
aenum==3.1.15
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
Deprecated==1.2.14
Flask==2.2.2
frozenlist==1.4.0
future==0.18.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==3.1.0
MarkupSafe==2.1.1
multidict==6.0.4
pydantic==1.10.11
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.2.2
wrapt==1.15.0
yarl==1.9.2
